moved functions to util file
moved all study locus functions to separate util files
