+++ apps/genetics/src/config.js
@@ -1,6 +1,8 @@
 const config = {
+  //TODO: update api url before merging to main
   apiUrl:
-    window.configApiUrl ?? 'https://api.genetics.dev.opentargets.xyz/graphql',
+    // window.configApiUrl ?? 'https://api.genetics.dev.opentargets.xyz/graphql',
+    window.configApiUrl ?? 'https://api.genetics.opentargets.org/graphql', 
   googleTagManagerID: window.configGoogleTagManagerID ?? null,
   helpdeskEmail: window.configHelpdeskEmail ?? 'helpdesk@opentargets.org',
   profile: window.configProfile ?? {},
