+++ apps/genetics/src/utils/common.js
@@ -1,0 +1,10 @@
+
+/* 
+Example usage:
+const comparatorDiseaseName = generateComparator(d => d.disease.name);
+ */
+export const generateComparator = (accessor) => (a, b) => {
+    const aValue = accessor(a);
+    const bValue = accessor(b);
+    return aValue > bValue ? 1 : aValue === bValue ? 0 : -1;
+  };
