+++ apps/genetics/src/utils/index.js
@@ -6,18 +6,12 @@ export function sanitize(str) {
   return str.replace(/[^a-zA-Z0-9]/g, '');
 }
 
-/* 
-Example usage:
-const comparatorDiseaseName = generateComparator(d => d.disease.name);
- */
-export const generateComparator = (accessor) => (a, b) => {
-  const aValue = accessor(a);
-  const bValue = accessor(b);
-  return aValue > bValue ? 1 : aValue === bValue ? 0 : -1;
-};
 
 // Consants
 export const SIGNIFICANCE = -Math.log10(5e-8);
+
+// Common helpers
+export * from './common';
 
 // chromosome helpers
 export * from './chromosome';
